--- a/frameworks/Ruby/grape/grape.dockerfile
+++ b/frameworks/Ruby/grape/grape.dockerfile
@@ -10,4 +10,4 @@ RUN bundle install --jobs=4 --gemfile=/grape/Gemfile --path=/grape/grape/bundle
 
 EXPOSE 8080
 
-CMD bundle exec puma -t 8:32 -w 8 --preload -b tcp://0.0.0.0:8080 -e production
+CMD bundle exec puma -C config/puma.rb -b tcp://0.0.0.0:8080 -e production