## Before:

```
--------------------------------------------------------------------------------
VERIFYING QUERY COUNT FOR http://tfb-server:8080/query?queries=20
--------------------------------------------------------------------------------
Elapsed time:		        2.57 secs
Data transferred:	        0.31 MB
Response time:		        1.56 secs
Transaction rate:	      199.22 trans/sec
Throughput:		        0.12 MB/sec
Concurrency:		      311.40
Longest transaction:	        2.55
Shortest transaction:	        0.08

Auditing /FrameworkBenchmarks/frameworks/Ruby/grape:
grape: Verify time: 55s

Auditing /FrameworkBenchmarks/frameworks/Ruby/grape:
grape-unicorn: Verify time: 55s
```
https://github.com/TechEmpower/FrameworkBenchmarks/actions/runs/6381783125/job/17318994384

## After:

```
--------------------------------------------------------------------------------
VERIFYING QUERY COUNT FOR http://tfb-server:8080/query?queries=20
--------------------------------------------------------------------------------
Elapsed time:		        1.18 secs
Data transferred:	        0.31 MB
Response time:		        0.71 secs
Transaction rate:	      433.90 trans/sec
Throughput:		        0.26 MB/sec
Concurrency:		      308.08
Longest transaction:	        1.17
Shortest transaction:	        0.01

Auditing /FrameworkBenchmarks/frameworks/Ruby/grape:
grape: Verify time: 48s

Auditing /FrameworkBenchmarks/frameworks/Ruby/grape:
grape-unicorn: Verify time: 52s
```
https://github.com/TechEmpower/FrameworkBenchmarks/actions/runs/6817552651/job/18541298655?pr=8534